Liberation Day
Each holoboard (except the first one) now has at least 2 marines at it. This is increased to 3 on hard and above.
There is now a viking in the square on hard, or a firebat on brutal
The two statues in the logistics compound now have 2 marines, 3 on hard and above
The third viking at the logistics thing is now also present on hard.
The firebat at the logistics thing is now also present on hard.
Moved the marines after the first holoboard back, added a group between them and the holoboard.
The square now has 13 marines. On hard, 3 of them are replaced by a viking. On brutal, it is a firebat instead.
There is now 4 marines on the path to the APC from the square. This is increased to 6 on hard, or a viking and 2 marines on Brutal
The first group of marines on the road to the town now has 5 marines.
Killing 3 and 6 holoboards rewards you with an additional marine squad.
Killing all 6 holoboards gives the Infantry Ultra Capacitors upgrade (+20% infantry attack speed)