Trains, Bel'shir
Trains:
You no longer start with a refinery
Removed one of the starting supply depots

Bel'shir:
The protoss now warp in guards for the 5 spread nodes. 1 warp prism, 2 stalkers, 1 scout, and 2 void rays, plus 1 colossus on hard, and 1 HT on brutal. These units have bullies, and are replaced as often as every 2.5 minutes on brutal.
Relocated the node right outside your base a bit to the west so the colossus there can't hit you right away.
The expansion now has a HT on brutal
Removed one of the nodes at the expo.
Added a node near the top-right bonus. It also starts empty.