Bel'shir
Bel'shir:
Upgraded the cyber core at the west protoss outpost to a nexus.
Added a cyber core to the center of the map.
Removed a void ray and stalker from the west outpost
Upgraded the Corner node to a Medium
Upgraded the Terrazine node to a Medium
The Node Defenders no longer get Colossi and HT right away, instead happening on the third and fourth reinforcement wave respectively
The Node Defenders are no longer warped in right away, instead waiting until the Warp Prism is in place.
High Templar energy is now only replenished for Node Defenders for Player 3, not all HT on the map.
Increased the minimum units each group of Node Defenders will maintain
Each area has some "stickyness" with the bullies. They won't give up the area right away.
There is now a dynamic number of warp prisms warped in.
Fixed the bullies on stuff in the top-right being disabled immediately
Fixed more warp-ins happening because the check area was too small
Node Defenders no longer count structures as units for the purposes of warping in reinforcements