Zero Hour, Outlaws, The Dig
Zero Hour:
Moved both of the low ground spines so they don't block the Omegalisks.

Outlaws:
Added a highlight when you build a tech lab, and text for what it does.

The Dig:
The Protoss attack waves now use a repeating scaling system that may or may not be broken.

Cutthroat:
Orlan's SCVs now have to "unload"
Contract price is 80k again