The Dig
Edited the Fast Build behaviour to affect charges and cooldowns, so it works on Warp Gates.
The Laser Drill now takes 200 power

The Dig:
Added 6 depots to start, to power the laser drill
Added a sensor tower next to the drill, for build radius
Slightly adjusted the high ground at the north ramp
More units in the protoss bases to feed attack waves
Adjusted Carrier, Void Ray, and Scout equations.
The laser drill is now owned by you, but is not commandable until Matt sees archons.