The Dig
The Dig:
Removed the base Immortals that always come.
Changed the equation for how many immortals come.
Made the large node full to start.
Made the 2 small main nodes empty to start.
Added a second large node in the main to regen money faster
Replaceed the medium node at the expansion with a large node and a small node