Bel'shir
The protoss no longer warp in extra warp prisms.
Counterattacks are no longer tied to killing warp prisms specifically
Counterattack strength is now tied to how much of the area attacked was cleared out.
Relocated the first relic to near the 2 east altars, so it's not so immediate that you get it.